+++ clang/docs/ReleaseNotes.rst
@@ -1250,6 +1250,7 @@ OpenMP Support
 - Added support for 'omp assume' directive.
 - Added support for 'omp scope' directive.
 - Added support for allocator-modifier in 'allocate' clause.
+- Added support for 'omp dispatch' directive.
 
 Improvements
 ^^^^^^^^^^^^